p/movil_1/requirements.txt
allure-behave==2.13.5
allure-pytest==2.13.5
allure-python-commons==2.13.5
Appium-Python-Client==2.11.1
attrs==25.1.0
behave==1.2.6
certifi==2025.1.31
h11==0.14.0
idna==3.10
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==24.2
parse==1.20.2
parse_type==0.6.4
pluggy==1.5.0
PySocks==1.7.1
pytest==8.3.4
selenium==4.29.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
